Remove mock updater
