+++ internal/test/test.go
@@ -38,3 +38,8 @@ func CmpSliceBytes(a, b []json.RawMessage) bool {
 	}
 	return true
 }
+
+type closingError struct{}
+
+func (e closingError) Closing()      {}
+func (e closingError) Error() string { return "closing" }
